+++ src/main/java/navik/domain/board/controller/BoardController.java
@@ -18,8 +18,9 @@ import navik.domain.board.dto.BoardCreateDTO;
 import navik.domain.board.dto.BoardLikeDTO;
 import navik.domain.board.dto.BoardResponseDTO;
 import navik.domain.board.dto.BoardUpdateDTO;
+import navik.domain.board.service.BoardCommandService;
 import navik.domain.board.service.BoardLikeService;
-import navik.domain.board.service.BoardService;
+import navik.domain.board.service.BoardQueryService;
 import navik.global.apiPayload.ApiResponse;
 import navik.global.apiPayload.code.status.GeneralSuccessCode;
 import navik.global.auth.annotation.AuthUser;
@@ -29,8 +30,9 @@ import navik.global.dto.CursorResponseDto;
 @RequestMapping("/v1/boards")
 @RequiredArgsConstructor
 public class BoardController implements BoardControllerDocs {
-	private final BoardService boardService;
+	private final BoardQueryService boardQueryService;
 	private final BoardLikeService boardLikeService;
+	private final BoardCommandService boardCommandService;
 
 	/**
 	 * 게시글 전체 조회
@@ -44,7 +46,7 @@ public class BoardController implements BoardControllerDocs {
 		@RequestParam(value = "cursor", required = false) Long lastId,
 		@RequestParam(value = "size", defaultValue = "10") int pageSize
 	) {
-		CursorResponseDto<BoardResponseDTO.BoardDTO> response = boardService.getBoardList(lastId, pageSize);
+		CursorResponseDto<BoardResponseDTO.BoardDTO> response = boardQueryService.getBoardList(lastId, pageSize);
 		return ApiResponse.onSuccess(GeneralSuccessCode._OK, response);
 	}
 
@@ -62,7 +64,7 @@ public class BoardController implements BoardControllerDocs {
 		@RequestParam(value = "cursor", required = false) Long lastId,
 		@RequestParam(value = "size", defaultValue = "10") int pageSize
 	) {
-		CursorResponseDto<BoardResponseDTO.BoardDTO> response = boardService.getBoardListByJob(jobName, lastId,
+		CursorResponseDto<BoardResponseDTO.BoardDTO> response = boardQueryService.getBoardListByJob(jobName, lastId,
 			pageSize);
 		return ApiResponse.onSuccess(GeneralSuccessCode._OK, response);
 	}
@@ -78,7 +80,7 @@ public class BoardController implements BoardControllerDocs {
 		@RequestParam(value = "cursor", required = false) String cursor,
 		@PageableDefault(size = 10) Pageable pageable
 	) {
-		BoardResponseDTO.HotBoardListDTO response = boardService.getHotBoardList(cursor, pageable);
+		BoardResponseDTO.HotBoardListDTO response = boardQueryService.getHotBoardList(cursor, pageable);
 		CursorResponseDto<BoardResponseDTO.BoardDTO> result = CursorResponseDto.of(
 			response.getBoardList(),
 			response.getHasNext(),
@@ -102,7 +104,7 @@ public class BoardController implements BoardControllerDocs {
 		@RequestParam(value = "cursor", required = false) Long lastId,
 		@RequestParam(value = "size", defaultValue = "10") int size
 	) {
-		CursorResponseDto<BoardResponseDTO.BoardDTO> response = boardService.searchBoard(keyword, lastId, size);
+		CursorResponseDto<BoardResponseDTO.BoardDTO> response = boardQueryService.searchBoard(keyword, lastId, size);
 		return ApiResponse.onSuccess(GeneralSuccessCode._OK, response);
 	}
 
@@ -115,7 +117,7 @@ public class BoardController implements BoardControllerDocs {
 	public ApiResponse<BoardResponseDTO.BoardDTO> getBoardDetail(
 		@PathVariable Long boardId
 	) {
-		BoardResponseDTO.BoardDTO boardDetail = boardService.getBoardDetail(boardId);
+		BoardResponseDTO.BoardDTO boardDetail = boardQueryService.getBoardDetail(boardId);
 		return ApiResponse.onSuccess(GeneralSuccessCode._OK, boardDetail);
 	}
 
@@ -131,7 +133,7 @@ public class BoardController implements BoardControllerDocs {
 		@RequestBody @Valid BoardCreateDTO request,
 		@AuthUser Long userId
 	) {
-		Long boardId = boardService.createBoard(userId, request);
+		Long boardId = boardCommandService.createBoard(userId, request);
 		return ApiResponse.onSuccess(GeneralSuccessCode._OK, boardId);
 	}
 
@@ -148,7 +150,7 @@ public class BoardController implements BoardControllerDocs {
 		@RequestBody @Valid BoardUpdateDTO request,
 		@AuthUser Long userId
 	) {
-		Long updatedBoardId = boardService.updateBoard(boardId, userId, request);
+		Long updatedBoardId = boardCommandService.updateBoard(boardId, userId, request);
 		return ApiResponse.onSuccess(GeneralSuccessCode._OK, updatedBoardId);
 	}
 
@@ -163,7 +165,7 @@ public class BoardController implements BoardControllerDocs {
 		@PathVariable Long boardId,
 		@AuthUser Long userId
 	) {
-		boardService.deleteBoard(boardId, userId);
+		boardCommandService.deleteBoard(boardId, userId);
 		return ApiResponse.onSuccess(GeneralSuccessCode._DELETED);
 	}
