+++ src/main/java/navik/domain/board/service/BoardQueryService.java
@@ -12,14 +12,11 @@ import org.springframework.transaction.annotation.Transactional;
 
 import lombok.RequiredArgsConstructor;
 import navik.domain.board.converter.BoardConverter;
-import navik.domain.board.dto.BoardCreateDTO;
 import navik.domain.board.dto.BoardResponseDTO;
-import navik.domain.board.dto.BoardUpdateDTO;
 import navik.domain.board.entity.Board;
 import navik.domain.board.repository.BoardLikeRepository;
 import navik.domain.board.repository.BoardRepository;
 import navik.domain.board.repository.CommentRepository;
-import navik.domain.users.entity.User;
 import navik.domain.users.repository.UserRepository;
 import navik.global.apiPayload.code.status.GeneralErrorCode;
 import navik.global.apiPayload.exception.handler.GeneralExceptionHandler;
@@ -27,7 +24,7 @@ import navik.global.dto.CursorResponseDto;
 
 @Service
 @RequiredArgsConstructor
-public class BoardService {
+public class BoardQueryService {
 	private final UserRepository userRepository;
 	private final BoardRepository boardRepository;
 	private final BoardLikeRepository boardLikeRepository;
@@ -192,61 +189,5 @@ public class BoardService {
 			boardLikeRepository.countLikeByBoard(board),
 			commentRepository.countCommentByBoard(board)
 		);
-	}
-
-	/**
-	 * 게시글 생성
-	 * @param userId
-	 * @param request
-	 * @return
-	 */
-	@Transactional
-	public Long createBoard(Long userId, BoardCreateDTO request) {
-		User user = userRepository.findById(userId)
-			.orElseThrow(() -> new GeneralExceptionHandler(GeneralErrorCode.USER_NOT_FOUND));
-		Board board = Board.builder()
-			.user(user)
-			.articleTitle(request.getArticleTitle())
-			.articleContent(request.getArticleContent())
-			.articleViews(0)
-			.build();
-
-		return boardRepository.save(board).getId();
-	}
-
-	/**
-	 * 게시글 수정
-	 * @param boardId
-	 * @param userId
-	 * @param request
-	 * @return
-	 */
-	@Transactional
-	public Long updateBoard(Long boardId, Long userId, BoardUpdateDTO request) {
-		Board board = boardRepository.findById(boardId)
-			.orElseThrow(() -> new GeneralExceptionHandler(GeneralErrorCode.BOARD_NOT_FOUND));
-
-		if (!board.getUser().getId().equals(userId)) {
-			throw new GeneralExceptionHandler(GeneralErrorCode.AUTH_BOARD_NOT_WRITER);
-		}
-
-		board.updateBoard(request.getArticleTitle(), request.getArticleContent());
-		return boardRepository.save(board).getId();
-	}
-
-	/**
-	 * 게시글 삭제
-	 * @param boardId
-	 * @param userId
-	 */
-	@Transactional
-	public void deleteBoard(Long boardId, Long userId) {
-		Board board = boardRepository.findById(boardId) // 게시글 찾을 수 없음
-			.orElseThrow(() -> new GeneralExceptionHandler(GeneralErrorCode.BOARD_NOT_FOUND));
-
-		if (!board.getUser().getId().equals(userId)) { // 게시글 작성자가 아님
-			throw new GeneralExceptionHandler(GeneralErrorCode.AUTH_BOARD_NOT_WRITER);
-		}
-		boardRepository.delete(board);
 	}
 }
